# src/main/resources/openapi/country-api.yaml
openapi: 3.0.3
info:
  title: REST Countries API
  description: API for retrieving information about countries.
  version: 1.0.0
servers:
  - url: /rest
paths:
  /v2/all:
    get:
      operationId: getAllCountries
      summary: Get all countries
      tags:
        - country
      parameters:
        - $ref: '#/components/parameters/fieldsParam'
        - name: page
          in: query
          description: Page number (must be greater than 0)
          required: false
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          description: Number of results per page (must be > 0 and < 50)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50

      responses:
        '200':
          description: A list of all countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'

  /v2/name/{name}:
    get:
      operationId: getCountriesByName
      summary: Search countries by name
      tags:
        - country
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/fieldsParam'
      responses:
        '200':
          description: Countries matching the name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'

  /v2/alpha/{code}:
    get:
      operationId: getCountryByAlphaCode
      summary: Search country by ISO 3166-1 code
      tags:
        - country
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 3
        - $ref: '#/components/parameters/fieldsParam'
      responses:
        '200':
          description: A country matching the code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryResponse'

  /v2/alpha:
    get:
      operationId: getCountriesByAlphaCodes
      summary: Search countries by list of codes
      tags:
        - country
      parameters:
        - name: codes
          in: query
          required: true
          schema:
            type: string
          description: ISO codes separated by semicolons (e.g., `col;no;ee`)
        - $ref: '#/components/parameters/fieldsParam'
      responses:
        '200':
          description: Countries matching the codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'

  /v2/currency/{currency}:
    get:
      operationId: getCountriesByCurrency
      summary: Search countries by currency code
      tags:
        - country
      parameters:
        - name: currency
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/fieldsParam'
      responses:
        '200':
          description: Countries using the specified currency
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'

  /v2/lang/{language}:
    get:
      operationId: getCountriesByLanguage
      summary: Search countries by language code
      tags:
        - country
      parameters:
        - name: language
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/fieldsParam'
      responses:
        '200':
          description: Countries using the specified language
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'

  /v2/capital/{capital}:
    get:
      operationId: getCountriesByCapital
      summary: Search countries by capital city
      tags:
        - country
      parameters:
        - name: capital
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/fieldsParam'
      responses:
        '200':
          description: Countries matching the capital
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'

  /v2/callingcode/{callingCode}:
    get:
      operationId: getCountriesByCallingCode
      summary: Search countries by calling code
      tags:
        - country
      parameters:
        - name: callingCode
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/fieldsParam'
      responses:
        '200':
          description: Countries matching the calling code
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'

  /v2/region/{region}:
    get:
      operationId: getCountriesByRegion
      summary: Search countries by region
      tags:
        - country
      parameters:
        - name: region
          in: path
          required: true
          schema:
            type: string
            enum: [ Africa, Americas, Asia, Europe, Oceania ]
        - $ref: '#/components/parameters/fieldsParam'
      responses:
        '200':
          description: Countries in the specified region
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'

  /v2/regionalbloc/{regionalBloc}:
    get:
      operationId: getCountriesByRegionalBloc
      summary: Search countries by regional bloc
      tags:
        - country
      parameters:
        - name: regionalBloc
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/fieldsParam'
      responses:
        '200':
          description: Countries in the specified regional bloc
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'

  /v2/demonym/{demonym}:
    get:
      operationId: getCountriesByDemonym
      summary: Search countries by demonym
      tags:
        - country
      parameters:
        - name: demonym
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/fieldsParam'
      responses:
        '200':
          description: Countries with the specified demonym
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'

components:
  parameters:
    fieldsParam:
      name: fields
      in: query
      required: false
      schema:
        type: string
      description: >
        Filter the fields to include in the response (e.g. `name;capital;currencies`)

  schemas:
    CountryResponse:
      type: object
      properties:
        name:
          type: string
        topLevelDomain:
          type: array
          items: { type: string }
        alpha2Code:
          type: string
          minLength: 2
          maxLength: 3
        alpha3Code:
          type: string
        callingCodes:
          type: array
          items: { type: string }
        capital:
          type: string
        altSpellings:
          type: array
          items: { type: string }
        region:
          type: string
        subregion:
          type: string
        population:
          type: integer
        latlng:
          type: array
          items: { type: number }
        demonym:
          type: string
        area:
          type: number
        gini:
          type: number
        timezones:
          type: array
          items: { type: string }
        borders:
          type: array
          items: { type: string }
        nativeName:
          type: string
        numericCode:
          type: string
        currencies:
          type: array
          items:
            $ref: '#/components/schemas/CurrencyResponse'
        languages:
          type: array
          items:
            $ref: '#/components/schemas/LanguageResponse'
        translations:
          type: object
          additionalProperties:
            type: string
        flag:
          type: string
          format: uri
        regionalBlocs:
          type: array
          items:
            $ref: '#/components/schemas/RegionalBlocResponse'
        cioc:
          type: string
      required:
        - name
        - alpha2Code
        - alpha3Code
        - callingCodes
        - capital
        - region
        - subregion
        - population
        - latlng
        - demonym
        - area
        - timezones
        - borders
        - nativeName
        - numericCode
        - currencies
        - languages
        - translations
        - flag

    CurrencyResponse:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        symbol:
          type: string
      required:
        - code
        - name
        - symbol

    LanguageResponse:
      type: object
      properties:
        iso639_1:
          type: string
        iso639_2:
          type: string
        name:
          type: string
        nativeName:
          type: string
      required:
        - iso639_1
        - iso639_2
        - name
        - nativeName

    RegionalBlocResponse:
      type: object
      properties:
        acronym:
          type: string
        name:
          type: string
        otherAcronyms:
          type: array
          items: { type: string }
        otherNames:
          type: array
          items: { type: string }
      required:
        - acronym
        - name
        - otherAcronyms
        - otherNames